backend/requirements.txt
fastapi
uvicorn[standard]
langchain-core
langchain-openai
langgraph
pillow
python-dotenv
httpx
pydantic
psycopg2-binary
